Add support for LB config in service config
